Allow COPYRIGHT environment variable to disable splash screen
Change COPYRIGHT variable assignment to use environment variable if set,
otherwise default to true. This allows ___UPDATE.sh to disable the
copyright splash by setting COPYRIGHT=false as an environment variable.

Using ${COPYRIGHT:-true} syntax ensures backward compatibility - if the
environment variable is not set, it defaults to true and shows the splash.